src/active_prm/eval/PRMBench/requirements.txt
accelerate==1.1.1
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
bitsandbytes==0.42.0
Brotli 
certifi 
charset-normalizer 
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
ffmpy==0.4.0
filelock 
fonttools==4.54.1
fsspec==2024.10.0
gmpy2 
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.26.2
idna
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2 
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
latex2mathml==3.77.0
loguru==0.7.2
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft 
mkl_random 
mpmath
narwhals==1.13.1
nest-asyncio==1.6.0
networkx 
numpy==1.26.4
openai==1.55.3
orjson==3.10.11
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow 
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks 
python-box==7.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML 
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests
rich==13.9.4
rpds-py==0.20.1
ruff==0.7.2
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.2
svgwrite==1.4.3
sympy 
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.20.1
tomlkit==0.12.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.46.0
triton==2.3.0
typer==0.12.5
typing_extensions 
tzdata==2024.2
urllib3 
uvicorn==0.32.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
